requirement.txt
python = 3.10.4
flask = *
panda = *
